fix：调整图表分析交互接口
调整图表分析交互接口
