+++ src/services/itc-platform/chartController.ts
@@ -2,12 +2,12 @@
 /* eslint-disable */
 import { request } from '@umijs/max';
 
-/** addChart POST /api/chart/add */
+/** addChart POST /api/bi/chart/add */
 export async function addChartUsingPost(
   body: API.ChartAddRequest,
   options?: { [key: string]: any },
 ) {
-  return request<API.BaseResponseLong_>('/api/chart/add', {
+  return request<API.BaseResponseLong_>('/api/bi/chart/add', {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
@@ -17,12 +17,12 @@ export async function addChartUsingPost(
   });
 }
 
-/** deleteChart POST /api/chart/delete */
+/** deleteChart POST /api/bi/chart/delete */
 export async function deleteChartUsingPost(
   body: API.DeleteRequest,
   options?: { [key: string]: any },
 ) {
-  return request<API.BaseResponseBoolean_>('/api/chart/delete', {
+  return request<API.BaseResponseBoolean_>('/api/bi/chart/delete', {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
@@ -32,12 +32,12 @@ export async function deleteChartUsingPost(
   });
 }
 
-/** editChart POST /api/chart/edit */
+/** editChart POST /api/bi/chart/edit */
 export async function editChartUsingPost(
   body: API.ChartEditRequest,
   options?: { [key: string]: any },
 ) {
-  return request<API.BaseResponseBoolean_>('/api/chart/edit', {
+  return request<API.BaseResponseBoolean_>('/api/bi/chart/edit', {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
@@ -47,136 +47,13 @@ export async function editChartUsingPost(
   });
 }
 
-/** genChartByAiAsync POST /api/chart/genChartByAiAsync */
-export async function genChartByAiAsyncUsingPost(
-  // 叠加生成的Param类型 (非body参数swagger默认没有生成对象)
-  params: API.genChartByAiAsyncUsingPOSTParams,
-  body: {},
-  file?: File,
-  options?: { [key: string]: any },
-) {
-  const formData = new FormData();
-
-  if (file) {
-    formData.append('file', file);
-  }
-
-  Object.keys(body).forEach((ele) => {
-    const item = (body as any)[ele];
-
-    if (item !== undefined && item !== null) {
-      if (typeof item === 'object' && !(item instanceof File)) {
-        if (item instanceof Array) {
-          item.forEach((f) => formData.append(ele, f || ''));
-        } else {
-          formData.append(ele, JSON.stringify(item));
-        }
-      } else {
-        formData.append(ele, item);
-      }
-    }
-  });
-
-  return request<API.BaseResponseBiResponse_>('/api/chart/genChartByAiAsync', {
-    method: 'POST',
-    params: {
-      ...params,
-    },
-    data: formData,
-    requestType: 'form',
-    ...(options || {}),
-  });
-}
-
-/** genChartByAiAsyncMq POST /api/chart/genChartByAiAsyncMq */
-export async function genChartByAiAsyncMqUsingPost(
-  // 叠加生成的Param类型 (非body参数swagger默认没有生成对象)
-  params: API.genChartByAiAsyncMqUsingPOSTParams,
-  body: {},
-  file?: File,
-  options?: { [key: string]: any },
-) {
-  const formData = new FormData();
-
-  if (file) {
-    formData.append('file', file);
-  }
-
-  Object.keys(body).forEach((ele) => {
-    const item = (body as any)[ele];
-
-    if (item !== undefined && item !== null) {
-      if (typeof item === 'object' && !(item instanceof File)) {
-        if (item instanceof Array) {
-          item.forEach((f) => formData.append(ele, f || ''));
-        } else {
-          formData.append(ele, JSON.stringify(item));
-        }
-      } else {
-        formData.append(ele, item);
-      }
-    }
-  });
-
-  return request<API.BaseResponseBiResponse_>('/api/chart/genChartByAiAsyncMq', {
-    method: 'POST',
-    params: {
-      ...params,
-    },
-    data: formData,
-    requestType: 'form',
-    ...(options || {}),
-  });
-}
-
-/** genChartByAiSync POST /api/chart/genChartByAiSync */
-export async function genChartByAiSyncUsingPost(
-  // 叠加生成的Param类型 (非body参数swagger默认没有生成对象)
-  params: API.genChartByAiSyncUsingPOSTParams,
-  body: {},
-  file?: File,
-  options?: { [key: string]: any },
-) {
-  const formData = new FormData();
-
-  if (file) {
-    formData.append('file', file);
-  }
-
-  Object.keys(body).forEach((ele) => {
-    const item = (body as any)[ele];
-
-    if (item !== undefined && item !== null) {
-      if (typeof item === 'object' && !(item instanceof File)) {
-        if (item instanceof Array) {
-          item.forEach((f) => formData.append(ele, f || ''));
-        } else {
-          formData.append(ele, JSON.stringify(item));
-        }
-      } else {
-        formData.append(ele, item);
-      }
-    }
-  });
-
-  return request<API.BaseResponseBiResponse_>('/api/chart/genChartByAiSync', {
-    method: 'POST',
-    params: {
-      ...params,
-    },
-    data: formData,
-    requestType: 'form',
-    ...(options || {}),
-  });
-}
-
-/** getChartVOById GET /api/chart/get */
+/** getChartVOById GET /api/bi/chart/get */
 export async function getChartVoByIdUsingGet(
   // 叠加生成的Param类型 (非body参数swagger默认没有生成对象)
   params: API.getChartVOByIdUsingGETParams,
   options?: { [key: string]: any },
 ) {
-  return request<API.BaseResponseChart_>('/api/chart/get', {
+  return request<API.BaseResponseChart_>('/api/bi/chart/get', {
     method: 'GET',
     params: {
       ...params,
@@ -185,12 +62,12 @@ export async function getChartVoByIdUsingGet(
   });
 }
 
-/** listChartByPage POST /api/chart/list/page */
+/** listChartByPage POST /api/bi/chart/list/page */
 export async function listChartByPageUsingPost(
   body: API.ChartQueryRequest,
   options?: { [key: string]: any },
 ) {
-  return request<API.BaseResponsePageChart_>('/api/chart/list/page', {
+  return request<API.BaseResponsePageChart_>('/api/bi/chart/list/page', {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
@@ -200,12 +77,12 @@ export async function listChartByPageUsingPost(
   });
 }
 
-/** listMyChartByPage POST /api/chart/my/list/page */
+/** listMyChartByPage POST /api/bi/chart/my/list/page */
 export async function listMyChartByPageUsingPost(
   body: API.ChartQueryRequest,
   options?: { [key: string]: any },
 ) {
-  return request<API.BaseResponsePageChart_>('/api/chart/my/list/page', {
+  return request<API.BaseResponsePageChart_>('/api/bi/chart/my/list/page', {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
@@ -215,12 +92,12 @@ export async function listMyChartByPageUsingPost(
   });
 }
 
-/** updateChart POST /api/chart/update */
+/** updateChart POST /api/bi/chart/update */
 export async function updateChartUsingPost(
   body: API.ChartUpdateRequest,
   options?: { [key: string]: any },
 ) {
-  return request<API.BaseResponseBoolean_>('/api/chart/update', {
+  return request<API.BaseResponseBoolean_>('/api/bi/chart/update', {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
